refactor: extract RateLimiter to pkg/rate_limiter package
Move RateLimiter interface, UnlimitedRateLimiter, MaximumRateLimiter,
and NewRateLimiter constructor from modes.go into a new dedicated
package pkg/rate_limiter.

This is a pure code move with no behavior change. All callers in
modes.go, main.go, modes_test.go, and integration_test.go are updated
to use the rate_limiter package prefix.
